chore(ci): add hassfest job in GitHub Actions workflow
Introduce a new hassfest job to the GitHub Actions
workflow. This change aims to leverage Home Assistant's
hassfest action for validating the setup and adherence
to their best practices. Including this job ensures that
any custom integrations maintain compatibility with Home
Assistant standards and guidelines, enhancing the quality
and reliability of the integration.